Inline vortex_arrow::export_array as PyArray.to_arrow
A bit of a nit, but: the indirection confused me more than helped
me navigate the code-base.

Moreover, inlining the code revealed the uses of unwrap which now
return errors to Python instead of panic'ing.
