+++ pyvortex/src/array.rs
@@ -1,5 +1,9 @@
+use arrow::array::{Array as ArrowArray, ArrayRef};
+use arrow::pyarrow::ToPyArrow;
 use paste::paste;
+use pyo3::exceptions::PyValueError;
 use pyo3::prelude::*;
+use pyo3::types::{IntoPyDict, PyList};
 use vortex::array::{
     Bool, BoolArray, BoolEncoding, Chunked, ChunkedArray, ChunkedEncoding, Constant, ConstantArray,
     ConstantEncoding, Primitive, PrimitiveArray, PrimitiveEncoding, Sparse, SparseArray,
@@ -8,7 +12,7 @@ use vortex::array::{
 };
 use vortex::compute::take;
 use vortex::encoding::EncodingRef;
-use vortex::{Array, ArrayDType, ArrayData, ArrayDef, ToArray};
+use vortex::{Array, ArrayDType, ArrayData, ArrayDef, IntoCanonical, ToArray};
 use vortex_alp::{ALPArray, ALPEncoding, ALP};
 use vortex_dict::{Dict, DictArray, DictEncoding};
 use vortex_fastlanes::{
@@ -24,7 +28,6 @@ use vortex_zigzag::{ZigZag, ZigZagArray, ZigZagEncoding};
 
 use crate::dtype::PyDType;
 use crate::error::PyVortexError;
-use crate::vortex_arrow;
 
 #[pyclass(name = "Array", module = "vortex", sequence, subclass)]
 pub struct PyArray {
@@ -179,7 +182,51 @@ impl PyArray {
 #[pymethods]
 impl PyArray {
     fn to_arrow(self_: PyRef<'_, Self>) -> PyResult<Bound<PyAny>> {
-        vortex_arrow::export_array(self_.py(), &self_.inner)
+        // NOTE(ngates): for struct arrays, we could also return a RecordBatchStreamReader.
+        // NOTE(robert): Return RecordBatchStreamReader always?
+        let py = self_.py();
+        let vortex = &self_.inner;
+
+        let chunks: Vec<ArrayRef> = if let Ok(chunked_array) = ChunkedArray::try_from(vortex) {
+            chunked_array
+                .chunks()
+                .map(|chunk| -> PyResult<ArrayRef> {
+                    Ok(chunk
+                        .into_canonical()
+                        .map_err(PyVortexError::map_err)?
+                        .into_arrow())
+                })
+                .collect::<PyResult<Vec<ArrayRef>>>()?
+        } else {
+            vec![vortex
+                .clone()
+                .into_canonical()
+                .map_err(PyVortexError::map_err)?
+                .into_arrow()]
+        };
+        if chunks.is_empty() {
+            return Err(PyValueError::new_err("No chunks in array"));
+        }
+
+        // Export the schema once
+        let data_type = chunks[0].data_type().clone();
+        let pa_data_type = data_type.to_pyarrow(py)?;
+
+        // Iterate each chunk, export it to Arrow FFI, then import as a pyarrow array
+        let chunks: PyResult<Vec<PyObject>> = chunks
+            .iter()
+            .map(|arrow_array| arrow_array.into_data().to_pyarrow(py))
+            .collect();
+
+        // Import pyarrow and its Array class
+        let mod_pyarrow = PyModule::import_bound(py, "pyarrow")?;
+
+        // Combine into a chunked array
+        mod_pyarrow.call_method(
+            "chunked_array",
+            (PyList::new_bound(py, chunks?),),
+            Some(&[("type", pa_data_type)].into_py_dict_bound(py)),
+        )
     }
 
     fn __len__(&self) -> usize {
